Update 1.01.02-whatIsIT.adoc
Minor grammatical and typographical updates, along with a non-minor proposed integration of the IT function vs problem domain sidebar into the text. 

I consider a sidebar something that is interesting yet ultimately not critical to understanding the main material. This discussion is far too important for that because of the contentiousness of discussions that you mention.This also make the Carr discussion easier to absorb as it can be linked to this material.